+++ src/intentatlas/adapters/__init__.py
@@ -1,0 +1,14 @@
+from .base import AdapterContext, GraphFragment, LanguageAdapter
+from .javascript import JavaScriptAdapter
+from .python import PythonAdapter
+
+BUILTIN_ADAPTERS: tuple[LanguageAdapter, ...] = (PythonAdapter(), JavaScriptAdapter())
+
+__all__ = [
+    "BUILTIN_ADAPTERS",
+    "AdapterContext",
+    "GraphFragment",
+    "JavaScriptAdapter",
+    "LanguageAdapter",
+    "PythonAdapter",
+]